setting up .irssi repo now

+++ setup.fish
@@ -32,29 +32,30 @@ if not test (type nvim)
   curl -sLf https://spacevim.org/install.sh | bash
 end
 
-# set up fonts
-mkdir -p /tmp/adodefont
-pushd /tmp/adodefont
-wget -q --show-progress -O source-code-pro.zip https://github.com/adobe-fonts/source-code-pro/archive/2.030R-ro/1.050R-it.zip
-unzip -q source-code-pro.zip -d source-code-pro
-mkdir -p ~/.fonts
-cp -v source-code-pro/*/OTF/*.otf ~/.fonts/
-fc-cache -f
-rm -rf source-code-pro{,.zip}
-popd
-# wget https://github.com/adobe-fonts/source-code-pro/archive/2.030R-ro/1.050R-it.zip
-# unzip 1.050R-it.zip
-# if not test -d ~/.fonts
-  # mkdir -p ~/.fonts
-# end
-# cp source-code-pro-*-it/OTF/*.otf ~/.fonts/
-# rm -rf source-code-pro-2.030R-ro-1.050R-it/
-# fc-cache -f -v
+# set up fonts if necessary
+if not test -e ~/.fonts/SourceCodePro-Semibold.otf
+  mkdir -p /tmp/adodefont
+  pushd /tmp/adodefont
+  wget -q --show-progress -O source-code-pro.zip https://github.com/adobe-fonts/source-code-pro/archive/2.030R-ro/1.050R-it.zip
+  unzip -q source-code-pro.zip -d source-code-pro
+  mkdir -p ~/.fonts
+  cp -v source-code-pro/*/OTF/*.otf ~/.fonts/
+  fc-cache -f
+  rm -rf source-code-pro{,.zip}
+  popd
+end
+
 # ensure my spacevim repo is cloned
 if not test -d ~/.SpaceVim.d
   if not git clone git@bitbucket.org:jhessin/vim.git ~/.SpaceVim.d
     echo "git clone failed - ensure you have your key set up with bitbucket"
   end
+end
+
+# ensure my .irssi repo is cloned
+if not test -d ~/.irssi/.git
+  rm -rf ~/.irssi
+  hub clone jhessin/.irssi ~/.irssi
 end
 
 update
